@@ -1,8 +1,3 @@
-/**
- * Created by Mihail on 12/1/2016.
- */
-'use strict';
-
 import React from 'react';
 
 class Wrap extends React.Component {
@@ -10,4 +5,5 @@ class Wrap extends React.Component {
 		return <div>{this.props.children}</div>;
 	}
 }
+
 export default Wrap;
