@@ -1,8 +1,3 @@
-/**
- * Created by Mihail on 12/6/2016.
- */
-'use strict';
-
 const EventsConsts = {
 	'USER_EXISTS': 'userExists',
 	'CHANGE': 'change',
@@ -13,4 +8,4 @@ const EventsConsts = {
 	'LOGOUT': 'logout'
 };
 
-export default EventsConsts
+export default EventsConsts;
