@@ -1,8 +1,3 @@
-/**
- * Created by Mihail on 11/30/2016.
- */
-'use strict';
-
 import React from 'react';
 
 class Logo extends React.Component {
@@ -10,4 +5,4 @@ class Logo extends React.Component {
 		return <div id="logo" />;
 	}
 }
-export default Logo
+export default Logo;
